Refactored the Noark 5.3 package structure model
The necessity to store more information about a single Noark 5.3
extraction package entity required that the way we represent it in the
model is refactored. This commit introduces the Noark53PackageEntity
that represents a single Noark 5.3 XML file, its corresponding XSD
schemas (both Noark-distributed and contained in the extraction
package), and a boolean value indicating whether it is mandatory. This
class and the Noark53PackageStructure were extracted as stand-alone
classes.

In addition to the above refactoring, loependeJournal.xml and
offentligJournal.xml are now considered optional in the Noark 5.3
extraction package.

Fixes issue #2.