Fix duplicate System IDs in Extraction-wide System ID Uniqueness test
Previously, if a Classification system was linked to more than one
Series, the validator would erroneously count the system IDs of that
Classification system and its Classes as many times as there were
Series connected to it.

The issue stemmed from the fact that arkivstruktur.xml may list a
Classification system multiple times if multiple Series refer to it.
During parsing of arkivstruktur.xml, however, we have no means of
identifying whether we have encountered that specific Classification
System or not. As a result, we end up persisting the Classification
System and its Classes multiple times.

To remedy the above, we now aggregate the Classification System and
Class System IDs when performing the Extraction-wide System ID
Uniqueness test.

Fixes issue #4.