Refactor the ExcelReport
The previous implementation of the ExcelReport generation logic was
relatively difficult to follow due to the multiple links between
methods and unclear naming. There is still room for improvement with
the new implemenation, but the ExcelReport generation logic is now much
more easily readable.

Added JavaDoc describing the purpose and target output for each sheet
in the generated report.