Add support for custom validation schemas
The Noark 5.3 extraction validation can now use externally-provided
Noark schemas during the validation process. Any incompatibilities
between the XML files in the extraction package and the custom schemas
will be reported as errors with schema type CUSTOM.

The new «-custom-schema-location» command parameter expects a directory
containing zero or more Noark schemas. Users need not provide all
required schemas in that directory, but if they do, they must use the
Noark 5.3 schema names. If a required schema is not found
in the directory, its Noark 5.3 counterpart will be used instead.

Additionally, a user may include a UTF-8-encoded description.txt file
in the aforementioned directory containing a summary of the changes
included in the specified custom schema files. The content of this file
will be included in the Execution Information section of all reports.

Any other files in the specified directory will simply be ignored.

Introduced exception message aggregators. The aggregators can be used
to group messages based on custom logic. The primary purpose of these
aggregators is to summarize the validation warnings and errors when
verifying the compliance of XML files to their corresponding XSD
schemas. The current implementation is strongly dependent on the
exception messages provided by the Xerces parser.