Add support for large excel reports generation
This is done using an SXSSFWorkbook which writes the contents
to the disk automatically once a certain threshold is reached.
The old XSSFWorkbook is replaced but can still be used with
the '-in-memory-xlsx-report' flag in case of any unforeseen
issues.

Also support for HSQL 'file' database is added which runs roughly
3 times slower than the 'in memory' one but can handle large
amounts of data. By default its contents are located in the OS
temp directory but can be explicitly specified.