Upgrade to Apache commons configuration 2 from Apache commons configuration
This required a small change to how we read the properties file in the
JAR resources.

Also, the list delimiter had to be explicitly set to "comma" as it was
not (after the change to 2.x) and the dokumentobjekt "_detected_*"
fields configuration was read as a single field instead of three.